added dedicated patcher to account for pure functions returning copy types
